fix(runtime): unwind journal gets its own TU; leak-gate hygiene in the matrix test
Round-2 CI found two consequences of the journal living inside
aether_panic.c:

Windows: the emitted C prologue references the track/forget symbols
from EVERY program, so the linker now pulled aether_panic.o from
libaether.a unconditionally, and with it the Windows DbgHelp
stack-trace imports. Hand-rolled link lines without -ldbghelp (the
aetherc_emit_main test, and any downstream consumer not using ae
cflags) broke. The journal now lives in runtime/actors/aether_unwind.c
with its own TLS frame depth; aether_panic.c calls
enter/exit/drain_current hooks, so programs without try/catch pull
only the journal object and never inherit the dbghelp requirement. The
header documents the link-surface constraint so a future edit does not
fold it back in.

macOS: the leaks(1) CI gate demands a fully clean exit; the unwind
matrix test now frees its map (and the escaped value it owns) before
reporting, 0 leaks under leaks --atExit.

Verified: aetherc_emit_main links and passes locally; MinGW
cross-compile of ae.c and aether_unwind.c clean (snprintf arg lists
rebalanced for the new manifest row); 229/229 unit tests; both panic
probes green; determinism and fmt gates green.
